% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_signature_enrichment.R
\name{geneset_enrichment}
\alias{geneset_enrichment}
\title{This is for geneset enrichment analysis}
\usage{
geneset_enrichment(geneset, LFC, LFC_name, figure_title)
}
\arguments{
\item{geneset}{user input}

\item{LFC}{CRISPR LFC value}

\item{LFC_name}{CRIPPR gene name relate with LFC}

\item{figure_title}{figure title}
}
\value{
enrichment matrix and enrichment plot
}
\description{
This is for geneset enrichment analysis
}
