% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankplot.R
\name{rankplot}
\alias{rankplot}
\title{rankplot for mageck}
\usage{
rankplot(gene_list, rra_LFC, rra_genename, cutoff = 1, figure_title)
}
\arguments{
\item{gene_list}{gene signature by user}

\item{rra_LFC}{LFC value of a RRA dataset}

\item{rra_genename}{gene name of a dataset}

\item{cutoff}{value for rank plot}
}
\description{
rankplot for mageck
}
