% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankplot_singlegene.R
\name{rankplot_singlegene}
\alias{rankplot_singlegene}
\title{rankplot of single gene for mageck}
\usage{
rankplot_singlegene(gene_list, rra_LFC, rra_genename, cutoff = 1, figure_title)
}
\arguments{
\item{gene_list}{gene signature by user(string)}

\item{rra_LFC}{LFC value of a RRA dataset}

\item{rra_genename}{gene name of a dataset}

\item{cutoff}{value for rank plot}
}
\description{
rankplot of single gene for mageck
}
