% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{singlegenebarplot}
\alias{singlegenebarplot}
\title{About project:This is for barplot}
\usage{
singlegenebarplot(
  inputdata,
  x,
  y,
  fill,
  sort.by.groups,
  ylab,
  xlab,
  legend.title,
  title
)
}
\arguments{
\item{inputdata}{This is a data frame}

\item{x}{x varibles for drawing (cohort name,character)}

\item{y}{y varibles for drawing (values,character)}

\item{fill}{fill colors (could be color name)}

\item{sort.by.groups}{logical value. If TRUE the data are sorted by groups. Used only when sort.val != "none".}

\item{ylab}{character vector specifying y axis labels. Use ylab = FALSE to hide ylab.}

\item{legend.title}{character vector for legend title}

\item{color}{ourline colors}

\item{sort.val}{a string specifying whether the value should be sorted. Allowed values are "none" (no sorting), "asc" (for ascending) or "desc" (for descending).}

\item{x.text.angle}{Rotate vertically x axis texts}

\item{rotate}{rotate angle}

\item{ggtheme}{function, ggplot2 theme name. Default value is theme_pubr(). Allowed values include ggplot2 official themes: theme_gray(), theme_bw(), theme_minimal(), theme_classic(), theme_void(), ....}
}
\description{
About project:This is for barplot
}
