% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stoufferzscore.R
\name{Stoufferzscore}
\alias{Stoufferzscore}
\title{This function are used for stouffer zscore}
\usage{
Stoufferzscore(w, z)
}
\arguments{
\item{w}{is the weight of the value}

\item{z}{is the value itself(the number of value is same as w)}
}
\value{
stoufferzscore value
}
\description{
This function are used for stouffer zscore
}
