% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanodisplay.R
\name{volcanodisplay}
\alias{volcanodisplay}
\title{This is for volcanodisplay}
\usage{
volcanodisplay(
  basic_path,
  specific_path,
  file_name,
  verified_gene,
  output_path
)
}
\arguments{
\item{basic_path}{your basic path for input file, string.}

\item{specific_path}{specific_path for your input now, string.}

\item{file_name}{file name you would like to input, string.}

\item{verified_gene}{the gene name vector which you would like to label.}

\item{output_path}{the output path which you would like to storage your figure.}
}
\value{
A figure of volcano
}
\description{
This is for volcanodisplay
}
