% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanodisplay_multigene.R
\name{volcanodisplaymultigene}
\alias{volcanodisplaymultigene}
\title{This is for volcanodisplay of single gene}
\usage{
volcanodisplaymultigene(file_path, verfied_gene, figure_name, output_path)
}
\arguments{
\item{file_path}{input file with absolute path.}

\item{verfied_gene}{the gene name vector which you would like to label.}

\item{figure_name}{the figure title and output figure name.}

\item{output_path}{the output path which you would like to storage your figure.}
}
\value{
A figure of volcano
}
\description{
This is for volcanodisplay of single gene
}
