% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanodisplayreplication.R
\name{volcanodisplayrep}
\alias{volcanodisplayrep}
\title{This is for volcano replication display}
\usage{
volcanodisplayrep(dd.rra, figure_title, verified_gene, output_path)
}
\arguments{
\item{dd.rra}{input dataframe which you have calculated the mean value.}

\item{figure_title}{figure title}

\item{verified_gene}{the gene name vector which you would like to label.}

\item{output_path}{the output path which you would like to storage your figure.}
}
\value{
A figure of volcano
}
\description{
This is for volcano replication display
}
